GridClear Optimization Update
Changed the GridClear function to customly delete all the game objects and set the undo deltas.
This is done to remove the redundancy of executing the code to update the SolidEdgeOutliner as there will be no blocks that have these edges when the grid is cleared.

Create a ForceClearGrid function to remove all undo stacks and act like a complete reset.
This will be used when loading a new file, as you shouldn't be able to undo into the old file.

Create the function ClearOperations to easily clear all the operations on the stack. This is used for the prior function.

Created a temp function for the file versioning in the future with todo comments in it.