Saving Multithreading
When a save is performed, the level grid is copied to a save grid buffer.
The save function creates a thread and saves that buffer data to a file. This allows for the level to be edited while saving.
However, no more saving can be done while the thread is still running.
Unity threads cannot run a lot of unity related systems, so any function call that needs to use them are put into a queue to be performed after the save thread is done.
This is mainly UI updates and StatusBar logging.