Move save threading to start on writeHelper
The save function checks how/if we can save and requires an overwrite check if saving to the same file. This can't be done in a thread and must be in the man thread.
Thus threading for saving is brought to the WriteHelper function where the writing is actually happening.