Converted versions to a struct, WIP diff function
Changed how versions work, now they are in a structure of auto and manual version ints. Manual always refers to the manual version we should be on, and auto is either 0 for referring to that manual, or > 0 for an auto save.

Some get version functions were combined.

Started work on getting the diffrences for saving to get the diffrences from a version instead of an old grid. This is because we could have an older version loaded and we need to load up to the newest version then diff that from our current grid.
Right now there won't be a need to get differences from an older version, but it could be used for flattening two versions more space efficiently, but would take more time.