Error logging and file deletion unmounting
Found error if a file was deleted and mounted, if you try to edit the grid it trys to save to the file that doesn't exist.
Fixed by unmounting the file if it is mounted.

Added error logging that exports error and exceptions to a txt in: "C:\Users\<user>\AppData\LocalLow\DigiPen Institute of Technology\Basic Level Builder" known as "Application.persistentDataPath"